+++ sdk/storage/azure-storage-file-share/CHANGELOG.md
@@ -1,5 +1,9 @@
 # Release History
 
+## Version XX.X.X (XXXX-XX-XX)
+- Added support for additional start copy parameters - FileSmbProperties, file permission, file permission copy mode, set archive and ignore read only
+
+
 ## Version 12.0.0 (2019-12-04)
 This package's
 [documentation](https://github.com/Azure/azure-sdk-for-java/blob/azure-storage-file_12.0.0/sdk/storage/azure-storage-file-share/README.md)